chore: remove cardano-cli ledger-state query, make AssetSupply.total optional
The benefit of this query is overshadowed by the resource demand. It's not a
supported interface, so was also always a risk for continual breaking changes.
The total ada supply will be restored if the information can be accessed via an
official API.

BREAKING CHANGE: AssetSupply.total is now an optional field, and will return null